Move `validateOpts()` to local_unix.go as it is not used on Windows
